backend/requirements.txt
aenum==3.1.15
affine==2.4.0
ai4bharat-transliteration==1.1.3
alabaster==1.0.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
bitarray==3.0.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contextily==1.6.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
dateparser==1.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
docutils==0.21.2
fairseq @ git+https://github.com/One-sixth/fairseq.git@44800430a728c2216fd1cf1e8daa672f50dfacba
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
fsspec==2024.12.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
hydra-core==1.3.2
idna==3.10
imagesize==1.4.1
indic-nlp-library==0.92
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
-e git+ssh://git@github.com/terrastack-tech/land_value.git@343689bd2808e97fc94560119ef02cf2c9cb8a41#egg=land_value&subdirectory=../../../../backend/submodules/land_value
lark==1.2.2
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
mercantile==1.2.1
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==3.1.1
progressbar2==4.5.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.0
pycparser==2.22
pydantic==1.10.20
pydload==1.0.9
pygeoapi==0.19.0
pygeofilter==0.3.1
pygeoif==1.5.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.3
razorpay==1.4.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
sacrebleu==2.5.1
sacremoses==0.1.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentinelhub==3.11.1
shapely==2.0.6
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
-e git+ssh://git@github.com/terrastack-tech/python_utils.git@57146ca13e3ba02ff0075e73798cad35fbd1ee7d#egg=terra_utils&subdirectory=../../../../../../../backend/submodules/land_value/submodules/python_utils
thin-plate-spline==1.1.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tinydb==4.8.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urduhack==0.1.4
urllib3==2.3.0
utm==0.7.0
Werkzeug==3.1.3
whitenoise==6.9.0
xyzservices==2024.9.0
zope.event==5.0
zope.interface==7.2
